requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

pypdf==4.3.1
numpy<2

sentence-transformers==3.0.1
faiss-cpu==1.7.4
rank-bm25==0.2.2

